Remove `flush` command
Flush command is overly complex/confusing and not really needed
as users can/should just use the `logout` command

Refs: #626
